backend/requirements.txt
fastapi
uvicorn
pymongo
qdrant-client
langgraph
langchain-core
langchain-google-genai
llama-parse
pydantic
python-dotenv
requests
cohere
google-generativeai
portalocker
numpy
dnspython
certifi
